Split code into separate modules. No functional changes.
